Don't Maintain State in Local Vars
Previously, the extension maintained relevant state in local variables in its
background page.  These pages are not persistent, even while the extension is
in use, and thus the state may be lost frequently. It seems a recent change in
Chrome causes this to happen much more often than previously. The solution is to
use the Chrome storage APIs to persistently store and recover the state so it is
not lost to background page garbage collection.