Add auto-update toggle and add papers buttons
Add two features that have been missing from PCS chair for some
time.  There is now an auto-update toggle, which controls whether
the site will get updated automatically when you visit a web page
or manually via a button on the page.

Additionally, buttons to add papers to the queue are automatically
added on the submissions page.