Fix table buttons
The previous code assumed a specific location for the ID column,
which was not always true, creating odd behavior.  It also was not
robust to changes introduced by users' reconfiguring their columns.
This new version searches for the ID column, doesn't display the
header if it doesn't add any buttons, and has a mutation observer
that monitors for heading changes and readds its header cell.